clauditor-24h.10: US-010 — extract remaining commands
Extract the five remaining command handlers out of cli/__init__.py,
finishing the US-008 / US-009 work so __init__.py becomes a thin
dispatcher.

Moved per-command modules (each with `add_parser(subparsers)` +
`cmd_<name>(args) -> int`):

- cli/triggers.py     (100 LOC)
- cli/extract.py      (157 LOC)
- cli/audit.py        (181 LOC)
- cli/suggest.py      (244 LOC; `cmd_suggest` + `_cmd_suggest_impl`)
- cli/compare.py      (396 LOC; `cmd_compare` + `_run_blind_compare`,
                      `_print_blind_report`, `_load_assertion_set`,
                      `_file_kind` — all single-consumer helpers
                      moved with the command)

cli/__init__.py: 1333 → 359 LOC. Remaining content is the argparse
dispatcher + shared helpers used by multiple commands
(`_unit_float`, `_positive_int`, `_append_validate_history`,
`_load_spec_or_report`, `_print_failing_transcript_slice`,
`_write_run_dir`, `_relative_to_repo`) plus `cmd_*` re-exports for
back-compat with test imports like
`from clauditor.cli import cmd_compare`.

Test patches for suggest moved from `clauditor.cli.propose_edits` /
`write_sidecar` / `load_suggest_input` to
`clauditor.cli.suggest.<symbol>`, following the US-008/US-009
pattern.

Help output: byte-identical for top-level and each of the five
commands (verified via diff against pre-change captures). All 1297
tests pass; coverage 96.57% (>=80% gate); ruff clean.
